# backend/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI小说创作系统",
    "description": "基于多Agent协作的智能小说创作平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/characters/": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "Create Character",
        "description": "创建角色",
        "operationId": "create_character_api_characters__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/characters/{character_id}": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "Get Character",
        "description": "获取角色详情",
        "operationId": "get_character_api_characters__character_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "角色管理"
        ],
        "summary": "Update Character",
        "description": "更新角色",
        "operationId": "update_character_api_characters__character_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "角色管理"
        ],
        "summary": "Delete Character",
        "description": "删除角色",
        "operationId": "delete_character_api_characters__character_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/characters/novel/{novel_id}": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "List Characters",
        "description": "获取小说的角色列表",
        "operationId": "list_characters_api_characters_novel__novel_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "importance_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "title": "Importance Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterResponse"
                  },
                  "title": "Response List Characters Api Characters Novel  Novel Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/characters/relationships": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "Create Character Relationship",
        "description": "创建角色关系",
        "operationId": "create_character_relationship_api_characters_relationships_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRelationshipCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterRelationshipResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/characters/{character_id}/relationships": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "Get Character Relationships",
        "description": "获取角色的所有关系",
        "operationId": "get_character_relationships_api_characters__character_id__relationships_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterRelationshipResponse"
                  },
                  "title": "Response Get Character Relationships Api Characters  Character Id  Relationships Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/characters/novel/{novel_id}/network": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "Get Character Network",
        "description": "获取小说的角色关系网络",
        "operationId": "get_character_network_api_characters_novel__novel_id__network_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "novel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Novel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterNetworkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/characters/appearances": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "Create Character Appearance",
        "description": "创建角色出场记录",
        "operationId": "create_character_appearance_api_characters_appearances_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterAppearanceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterAppearanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CharacterRelationshipCreate": {
        "properties": {
          "relationship_type": {
            "type": "string",
            "maxLength": 50,
            "title": "Relationship Type",
            "description": "关系类型"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "关系描述"
          },
          "strength": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Strength",
            "description": "关系强度",
            "default": 5
          },
          "development_stage": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Development Stage",
            "description": "发展阶段"
          },
          "established_in_chapter": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Established In Chapter",
            "description": "建立关系的章节"
          },
          "novel_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Novel Id",
            "description": "所属小说ID"
          },
          "character_a_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Character A Id",
            "description": "角色A的ID"
          },
          "character_b_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Character B Id",
            "description": "角色B的ID"
          }
        },
        "type": "object",
        "required": [
          "relationship_type",
          "novel_id",
          "character_a_id",
          "character_b_id"
        ],
        "title": "CharacterRelationshipCreate",
        "description": "创建角色关系请求"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}